+++ pkg/sdk/organization_accounts_impl_gen.go
@@ -2,9 +2,6 @@ package sdk
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"strings"
 )
 
 var _ OrganizationAccounts = (*organizationAccounts)(nil)
@@ -31,48 +28,6 @@ func (v *organizationAccounts) Show(ctx context.Context, request *ShowOrganizati
 	}
 	resultList := convertRows[organizationAccountDbRow, OrganizationAccount](dbRows)
 	return resultList, nil
-}
-
-// ShowParameters added manually
-func (v *organizationAccounts) ShowParameters(ctx context.Context) ([]*Parameter, error) {
-	return v.client.Parameters.ShowParameters(ctx, &ShowParametersOptions{
-		In: &ParametersIn{
-			Account: Bool(true),
-		},
-	})
-}
-
-// UnsetAllParameters added manually
-func (v *organizationAccounts) UnsetAllParameters(ctx context.Context) error {
-	return v.client.Accounts.UnsetAllParameters(ctx)
-}
-
-// UnsetPolicySafely added manually
-func (v *organizationAccounts) UnsetPolicySafely(ctx context.Context, kind PolicyKind) error {
-	unset := NewOrganizationAccountUnsetRequest()
-	switch kind {
-	case PolicyKindPasswordPolicy:
-		unset.WithPasswordPolicy(true)
-	case PolicyKindSessionPolicy:
-		unset.WithSessionPolicy(true)
-	default:
-		return fmt.Errorf("policy kind %s is not supported for account policies", kind)
-	}
-	err := v.client.OrganizationAccounts.Alter(ctx, NewAlterOrganizationAccountRequest().WithUnset(*unset))
-	// If the policy is not attached to the account, Snowflake returns an error.
-	if err != nil && strings.Contains(err.Error(), fmt.Sprintf("Any policy of kind %s is not attached to ACCOUNT", kind)) {
-		return nil
-	}
-	return err
-}
-
-// UnsetAll added manually
-func (v *organizationAccounts) UnsetAll(ctx context.Context) error {
-	return errors.Join(
-		v.client.OrganizationAccounts.UnsetAllParameters(ctx),
-		v.client.OrganizationAccounts.UnsetPolicySafely(ctx, PolicyKindPasswordPolicy),
-		v.client.OrganizationAccounts.UnsetPolicySafely(ctx, PolicyKindSessionPolicy),
-	)
 }
 
 func (r *CreateOrganizationAccountRequest) toOpts() *CreateOrganizationAccountOptions {
